+++ CHANGELOG.md
@@ -7,3 +7,6 @@
 ## 2026-02-10T01:13:34.069138+00:00
 - Workflows API endpoints: workflow-marketplace_patch_2026-02-10T01-13-25.603821+00-00.diff
 
+## 2026-02-10T01:31:42.703330+00:00
+- Workflows API endpoints: workflow-marketplace_patch_2026-02-10T01-31-33.777274+00-00.diff
+